Added Cell Spawning
Added Logic for cell spawning, the system calculates the left bottom point and the right top point from the GeoJSON data and then spawns a complete (rectangular) grid based on it. The Cells are attached to the Manager

Created C++ and BP actor for Grid Cell, to provide more flexibility for customizing the grid appearance in the future.